Re-work insertion into the priority stack.
There's a stack that maintains a queue of threads. If a new thread
arrives at the scheduler and the priority isn't in the stack then
we need to find the appropriate place to insert it.

Previously this would be an O(n2) problem as it iterated over the stack
using an index and a get() method to get the queue. However each get
call re-iterates over the linked list.

Now we can use an iterator but still iterator over to the list so the
insertion becomes O(n) instead. Will try and convert to binary search
later on.